Fix improper use of the three dots
In the previous version, the "trunc" var (i.e. the "..." string)
was added to the end of **every** truncated string, even though
in some edges cases, the string given as an argument does not
need to be truncated.
This commit first check if the given string needs truncation:
for instance, if the second argument 'num' is greater or equal
to the length of the first argument 'str', there's no need to
truncate.