Fix the console output after pushes to newArr
It appears that my previous solution was correct in terms of array
creation.
The problem lay in my misunderstanding of the construction of
console.log outputs.
If code reads:
    console.log("some string" + some_variable);
the output will be a **string**.
Whereas if code reads:
    console.log("some string", some_variable);
the output will be "some string" followed by the actual variable value.

So newArr is correctly created, I was just not displaying it correctly
in the console!