Test edge case: array.length not dividable by size
The current problem is that, when arr.length is not dividable by
size, the remaining items in the original arr are omitted from
newArr.

For instance:
    chunkArrayInGroups(["a", "b", "c", "d", "e"], 2)

will only give:
    [["a", "b"],["c","d"]]

as a result, omitting the last item "e".
I need to find a solution for this in the next commit(s).