Fix apparent previous "fail" (wasn't one!)
The problem lay in the fact that I renamed the function 'slasherFlick'
instead of 'slasher'.
This sounds as a dumb fail on the FCC part…
If I replace the function name in *both* the function declaration
and the call of the function, it should be OK.
My code **is** technically correct. It just fails the tests because
the test expect the function to be called 'slasher'…

Note to future self: DON'T CHANGE THE FUNCTION NAME in FCC placeholder code.